Age and Weight blocs fromJson casted as int

+++ lib/blocs/age_bloc/age_bloc.dart
@@ -14,7 +14,7 @@ class AgeBloc extends HydratedBloc<AgeEvent, int> {
 
   @override
   int? fromJson(Map<String, dynamic> json) {
-    return json['state'];
+    return json['state'] as int;
   }
 
   @override
